Edited template, updated pages all pages to Bootstrap and added contact form.
Edited template file to include a div for content. This div is a well
and has a slight offset on both sides.
All pages have been updated to use the new template, which means
everything looks nice and consistent.
I added a contact form. I made the form so it should hopefully work with
'send.php', but I haven't tested it yet.
I added 'contactForm.css' to 'css' for some custom form styles that I
didn't want to be on every page.
Edited 'stylesheet.css' to make all page headers be aligned to the
centre of the page.