Moved EEPS pages out of 'eeps' folder and deleted it
The EEPS pages were in their own 'eeps' folder. The problem with this is
that when you visited one of the EEPS pages, you were in a different
directory, so you couldn't get back to the main page. This could be
changed, but it's not worth. I moved the EEPS pages back into the main
directory and updated links to them.
Deleted 'eeps' folder, it is no longer needed.
Updated repository structure in README to reflect the deletion of the
'eeps' folder.